# feat: Add comprehensive performance optimizations for MCP TypeScript SDK