… src/shared/__tests__/performance-benchmark.test.ts
… src/shared/__tests__/performance-optimizations.test.ts
… src/shared/serialization.ts