@ihrpr 2025-07-14
Thank you for working on this! Given that the current SDK performance hasn't been a reported issue from users, and considering the maintenance overhead these changes would introduce, I don't think we should proceed with this PR at this time.

That said, there are great ideas here! If performance becomes a concern in the future, we could revisit specific optimizations in a more targeted way - perhaps focusing on one area at a time with clearer benchmarks against real-world usage patterns.